Move BCheck-related data to bcheck package.
