+++ src/main/java/data/bcheck/BCheck.java
@@ -1,4 +1,7 @@
-package data;
+package data.bcheck;
+
+import data.Item;
+import data.Tags;
 
 import java.io.IOException;
 import java.nio.file.Path;
