+++ src/main/java/repository/RepositoryFacadeFactory.java
@@ -2,8 +2,8 @@ package repository;
 
 import burp.api.montoya.http.Http;
 import client.GitHubClient;
-import data.BCheck;
-import data.BCheckFactory;
+import data.bcheck.BCheck;
+import data.bcheck.BCheckFactory;
 import file.finder.BCheckFileFinder;
 import file.temp.TempFileCreator;
 import file.zip.ZipExtractor;
