+++ src/test/java/data/BCheckFactoryTest.java
@@ -1,5 +1,7 @@
 package data;
 
+import data.bcheck.BCheck;
+import data.bcheck.BCheckFactory;
 import logging.Logger;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
